app-rutas - copia/src/requirements.txt
Flask>2.3.2
numpy>1.24.4
networkx>3.1
matplotlib>3.7.1
Pillow>9.5.0
mysql-connector-python>8.0.33

# Flask==3.0.3
# numpy==1.26.4
# networkx==3.3
# matplotlib==3.9.1
# pillow==10.4.0
# mysql-connector-python==9.0.0